practice-for-week-19-node-flask-microservices/flask_app/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Configuration:
    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
    # SQLALCHEMY_DATABASE_URI = os.environ.get("DATABASE_URL", "sqlite:///default.db")
    SQLALCHEMY_TRACK_MODIFICATIONS = False
